feat: enhance inventory and reporting features
- Added `formatQty` utility to format quantities with optional units.
- Updated `ItemTable`, `ProductWiseSalesTable`, and `TopItemsTable` to use `formatQty` for displaying quantities.
- Introduced delivery number input in `StoreIdentitySection` with saving functionality.
- Modified `BillingPage` to support item quantity entry by amount for items sold by weight.
- Enhanced billing store to manage item quantities and units more effectively.
- Updated types to include `soldByAmount` and `unit` for items and cart entries.
- Improved PDF generation for reports and receipts to include formatted quantities.
- Adjusted CSV report generation to round quantities to two decimal places.

+++ src-tauri/src/db/seed.rs
@@ -411,12 +411,12 @@ mod tests {
                 .map(|(_, qty)| qty)
                 .sum();
 
-            let stock: i64 = conn
+            let stock: f64 = conn
                 .query_row("SELECT stock_qty FROM items WHERE name = ?1", [name], |row| {
                     row.get(0)
                 })
                 .unwrap();
-            assert_eq!(stock, opening_stock - sold, "stock wrong for {}", name);
+            assert_eq!(stock, (opening_stock - sold) as f64, "stock wrong for {}", name);
         }
     }
